actionlint: run on every push
We currently run only when when workflow files are modified, but that
means we cannot make the results of the Zizmor scan a required status
check.

Let's just run this all the time so that we can make sure to never merge
regressions.

This will affect only one high-volume repo: Homebrew/cask. We can carve
out an exception for that if we need to. (Though I'm currently inclined
to do something similar on our high-volume repos anyway.)

See discussion at Homebrew/ci-orchestrator#255