workflows/actionlint: enable merge queue/group jobs.
This should allow us to make use of the GitHub merge queue to ensure
that we're not merging outdated code (or breaking `master`) but avoiding
the need to continually merge into/rebase PR branches.

This should be safe to merge as-is as is essentially a no-op without
the merge queue enabled.