actionlint: don't try upload SARIF on merge_group
From what I can see we aren't able to upload the SARIF record within the merge group because it can't find a ref to upload against.

This was seen in `homebrew/cask` - https://github.com/Homebrew/homebrew-cask/actions/runs/14210386041/job/39816439146

As far as I can see, no other repositories that this file is synced to have merge queue enabled, so the failure hasn't been seen until now.

---

An alternative may be to skip the upload job, and only run the `actionlint` side of the workflow when it is a merge_group, unless there is an alternative way to link to a ref.